passes example case, produces wrong answer
Current code produced the correct answer for the small example case.
However, the wrong answer is produced for the regular input.
The total pressure released by a valve if it is opened, is divided by
the square of the distance to the valve, to account for the travel time.
This approximates the travel time impact.